Use setuptools and encourage develop setup target
This is a library which is liable to change often, and using `develop`
instead of `install` is a much more convenient way to install (since
it makes a link to the code rather than copying it).

Additionally, remove some unneeded git ignore rules (since this code
should probably be run outside the repository itself).