Make mti-sources processing more strict
If --mti-sources is provided, then it MUST declare mti sources for
all masters, so don't silently fallback to automatic feature
generation.

In fact, this exposed a bug in fontmake.  When running against
NotoSansDevanagari sources, we now get:

>> Saving OTF for NotoSansDevanagari-Light
Traceback (most recent call last):
  File "/usr/local/bin/fontmake", line 9, in <module>
    load_entry_point('fontmake==1.0', 'console_scripts', 'fontmake')()
  File "build/bdist.linux-x86_64/egg/fontmake/__main__.py", line 42, in main
  File "build/bdist.linux-x86_64/egg/fontmake/font_project.py", line 221, in run_from_glyphs
  File "build/bdist.linux-x86_64/egg/fontmake/font_project.py", line 258, in run_from_ufos
  File "build/bdist.linux-x86_64/egg/fontmake/font_project.py", line 113, in build_otfs
  File "build/bdist.linux-x86_64/egg/fontmake/font_project.py", line 157, in save_otfs
KeyError: 'NotoSansDevanagari-Light'

That is because the sources call this font Thin, and the plist has that
and the mti-sources are called that.  But fontmake ignored the name
of the master and replaced it with Light, and is looking for Light.

That should be fixed separately.