Use logging module
This allows the user to specify whether they want to see the build
status messages or not via a --verbosity flag.

It has the side effect, however, of printing info from dependencies.
In the case of MutatorMath, which otherwise runs silently for a long
time, this isn't necessarily a bad thing. Though in some cases (e.g.
fontTools' subsetter) the info messages can be a bit noisy.