fontmake.main: allow to pass arguments to main()
When ArgumentParser.parse_args() method is passed 'None', it will read the argument from sys.argv as usual.
Otherwise, 'args' can be a list of strings, and is parsed the same way.

This can be useful if one wants to run fontmake from python code without running it in a subprocess (e.g. for writing tests)