Copy MTI feature files into master UFOs
After this change, the master UFOs are self-contained in the sense
that no external files are needed to compile them, even in workflows
that use Monotype (MTI) feature files instead of Adobe feature files.

We still need to change ufo2ft so it actually recognizes these newly
embedded files, but this can be done independently.

https://github.com/googlei18n/fontmake/issues/289