FDKFeatureCompiler: handle new makeotf with non-zero exit code
https://github.com/adobe-type-tools/afdko/issues/417

This way we support both the old and the new behaviour.

I also renamed the method from the deprecated setupFile_featureTables to buildTables
and removed the overridden __init__ as the 'mtiFeatures' argument is no longer used
by ufo2ft (one now needs to use a separate MTIFeatureCompiler class for those).