make zip containing fontmake.pyz and LICENSE; only build for 3.6 and 3.7
the zip contains a 'fontmake' (without .pyz extension) on Unix, and keeps the executable flag
so one doesn't need to rename and chmod +x fontmake before running, but just unzip it.

on windows we name it fontmake.pyz so that the py.exe lancher can parse the shebang and treat
it like a python script.

also, include a LICENSE file because.