read/apply ufo2ft 'featureWriters' setup if found in the designspace lib
upcoming glyphsLib will store a custom 'featureWriters' key in the generated
designspace <lib> element, so fontmake needs to read that and pass the option on
to the ufo2ft compile functions.

https://github.com/googlei18n/glyphsLib/pull/441

Following discussions in https://github.com/googlei18n/ufo2ft/issues/280, we decided
to revert the default mode for the ufo2ft kernFeatureWriter to skip (if feature already present).