move mutatormath to its own method; raise error with source layers for static instances
'sparse' masters (with outlines/metrics for subset of glyphs coming from a non-default layer)
are not supported by MutatorMath yet, so we need to prevent calling the latter with a designspace
that contains these, otherwise we would produce invalid output.
We shall either move to ufoProcessor (the successor of mutatorMath, which supports these layer-sources)
or something else that supports that.
For the time being, sparse masters are only supported for the variable font output, not for the static
fonts...