Fix generating instance UFOs
Since fontmake 1.9, attempting to build instance UFOs (-i -o ufo) from either -g (Glyphs) or -m (Designspace) inputs has been broken (i.e. doing nothing, leaving instance_ufo directory empty).

The problem was that the run_from_designspace method was splitting the list of outputs in two groups, static vs interpolatable, and calling the appropriate method for each; but at the same time it was ignoring the 'ufo' output altogether.
This fixes it, by assuming that any output which is not in the set of INTERPOLATABLE_OUTPUS is run through the 'static' part of the pipeline (the one that uses MutatorMath to interpolate static instances).