require glyphsLib 3.3.1
this is to blacklist glyphsLib 3.3.0, which changed the way width axis coordinates are interpreted (and thus the generated fvar and avar tables in variable fonts). glyphsLib v3.3.1 restores the previous behavior until we reach agreement in https://github.com/googlefonts/glyphsLib/issues/483.
The change in question is https://github.com/googlefonts/glyphsLib/pull/451