Enable writing 'public.skipExportGlyphs' list to DS/UFOs
When converting from .glyphs to DS/UFO, by default fontmake will store the
glyph export flags as 'public.skipExportGlyphs' list in the designspace and
UFO lib. This is used by ufo2ft to prune the glyphs set before the binary font
is built (decomposing any skipped glyphs that are used as components).

This behavior can be disabled with the option --no-write-skipexportglyphs,
which restore the original behavior whereby the glyph export flags are
written to the UFO glyphs' lib using the private com.schriftgestaltun.Glyphs.Export
boolean key, and the latter keys are used by fontmake to subset the generated
binary font after it has been built.