Copy font info attributes from default font to instance selectively
I overlooked that ufoProcessor does the same. Several attributes are actually specific to instances and should not be blindly inherited from the default font.

The ignore list is different from ufoProcessor. We additionally ignore:
- openTypeNameUniqueID
- openTypeNameWWSFamilyName
- openTypeNameWWSSubfamilyName
- openTypeOS2Panose

We copy over:
- familyName
- woffMajorVersion
- woffMetadataCopyright
- woffMetadataCredits
- woffMetadataDescription
- woffMetadataExtensions
- woffMetadataLicense
- woffMetadataLicensee
- woffMetadataTrademark
- woffMetadataVendor
- woffMinorVersion

This implies that the styleName attribute is now implicitly required. Not specifying it in the Designspace means it will not be set in the instance.