ignore glyphs in public.skipExportGlyphs when subsetting for 'Keep Glyphs' custom param
When UFO has both 'public.skipExportGlyphs' and the 'Keep Glyphs' custom parameter, then the final
glyph order in the built OTF will not include glyphs from skipExportGlyphs.
If we don't prune them when getting the UFO glyphOrder we may hit the assertion later on that
checks that the length of UFO and OTF glyphOrders is the same.