fix default ../instance_ufo dir relative to designspace dir
sources' and instances' `filename` attributes in designspace file are
required to be relative to the designspace directory.

Before https://github.com/googlefonts/fontmake/commit/471438025f7314627f30a1f1ceb21a7e31e08a9d, we used to call:

    instance_dir = os.path.relpath(instance_dir, designspace_dir)

unconditionally, whereas after that commit we changed to only relativize
instance_dir when it was provided as an absolute path, in order to avoid
a hard crash on Windows when relpath is called with two directories
located under different mount points.

The default instance directory in fontmake is traditionally called
'instance_ufo', and it's normally located at the same level as the
'master_ufo', where the .designspace file itself is located by default.

But because, ever since the above commit, we stopped calling relpath on
the default 'instance_ufo' directory (given it is not an absolute path),we ended up passing it unchanged to glyphsLib.to_designspace method; the latter prepends the instance_dir to the instances' filename attributes.

However, that means that we would end up with two 'instance_ufo' directories: one (produced upfront by os.mkdir) next to 'master_ufo', which is always left empty... The other one, _inside_ the directory where the designspace file is, which by default is inside 'master_ufo'.

This PR attempts to restore the old default behavior when outputting
instance UFOs: i.e. one gets a
master_ufo (containing master UFOs and the designspace file) and an instance_ufo located next to it.