don't save master/instance UFOs to disk unless requested with -o ufo
This should considerably speed up build time as we can now avoid serializing UFOs and build either VFs or static instances directly from in-memory objects.
A few places that assumed things were read from/written to filesystem had to be tweaked, but the normal behaviour is unchanged.
To save the UFOs you now have to explicitly include 'ufo' in the list of output formats.
And if the *only* output is -o ufo, then one can also use the --output-dir option (works for either masters or instances) or even the --output-path option when selecting to build one specific instance UFO.