add more test glyphs to IncompatibleSans
made the Bold 'space' glyph non-empty (the default 'Regular' is still empty) to trigger error when default is empty but non-default is not.

also added 'E' glyph which is non-empty in default 'Regular' master but it is empty in the non-default 'Bold' master. This should be allowed and pass the check, not raise an error. The empty non-default glyph should be treated as if it is not there.
This is to add support for sparse kerning.