ci: tell twine upload to --skip-existing
a minor issue in a setup.cfg file used at build time (which is not even included in the published .whl package) made it fail the upload of the v3.11.1 source distribution, however the wheel for 3.11.1 was published fine already.
If I want to retry uploading the fixed up sdist I need to tell twine to --skip-existing otherwise PyPI rejects the attempt to re-upload the wheel for 3.11.1 which has the same filename as the one already uploaded.
And really it is not worth cutting a 3.11.2 just for this.